created appRouter.js, pulled router logic into AppRouter to be imported to index.js. Moved history package and object to appRouter, moved auth0provider to appRouter, moved onRedirectCallback to appRouter. Working on logic for conditionally setting redirect URI to /profile if app_metadata.role || app_metadata.role !== 'null' || app_metadata.role !== 'undefined' or /plan if app_metadata.role === false || app_metadata.role === 'null' || app_metadata.role === 'undefined'
