+++ src/components/Profile/Profile.jsx
@@ -26,7 +26,7 @@ const Profile = ({ apiKey, setApiKey }) => {
       </Message.Content>
     </Message>
   )
-
+    
   const getApiKey = async () => {
     try {
       setKeyLoading(true)
