+++ src/index.jsx
@@ -3,28 +3,29 @@ import ReactDOM from 'react-dom'
 import App from './components/App'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import * as Sentry from '@sentry/browser'
-
 import * as serviceWorker from './serviceWorker'
+//import AppRouter
+import AppRouter from './appRouter'
 
 import './index.scss'
 import 'semantic-ui-css/semantic.min.css'
 import 'semantic-ui-css/semantic.min.css'
 
-import { Auth0Provider } from './hooks/useAuth0'
-import { Router } from 'react-router-dom'
-import history from './utils/history'
+// import { Auth0Provider } from './hooks/useAuth0'
+// import { Router } from 'react-router-dom'
+// import history from './utils/history'
 
 Sentry.init({
   dsn: 'https://ed2c8b82501542498bcdb20ad95bac85@sentry.io/1779492'
 })
 
-const onRedirectCallback = appState => {
-  history.push(
-    appState && appState.targetUrl
-      ? appState.targetUrl
-      : window.location.pathname
-  )
-}
+// const onRedirectCallback = appState => {
+//   history.push(
+//     appState && appState.targetUrl
+//       ? appState.targetUrl
+//       : window.location.pathname
+//   )
+// }
 
 const styleLink = document.createElement('link')
 styleLink.rel = 'stylesheet'
@@ -32,7 +33,15 @@ styleLink.href =
   'https://cdn.jsdelivr.net/npm/semantic-ui/dist/semantic.min.css'
 document.head.appendChild(styleLink)
 
+
+const jsx = (<AppRouter />)
+
 ReactDOM.render(
+  jsx,
+  document.getElementById('root')
+)
+
+/* ReactDOM.render(
   <Router history={history}>
     <Auth0Provider
       domain={process.env.REACT_APP_DOMAIN}
@@ -46,7 +55,7 @@ ReactDOM.render(
     </Auth0Provider>
   </Router>,
   document.getElementById('root')
-)
+) */
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
